# Rewrite Control.Effect.Interpret in terms of reflection