--- a/src/Control/Effect/Interpret.hs
+++ b/src/Control/Effect/Interpret.hs
@@ -4,6 +4,7 @@ module Control.Effect.Interpret
 ( runInterpret
 , runInterpretState
 , InterpretC(..)
+, Reifies
 ) where
 
 import Control.Applicative (Alternative(..))