--- a/src/Control/Effect/Interpret.hs
+++ b/src/Control/Effect/Interpret.hs
@@ -19,6 +19,7 @@ import Control.Monad.Trans.Class
 import Unsafe.Coerce (unsafeCoerce)
 
 
+-- | A @Handler@ is a function that interprets effects described by @sig@ into the carrier monad @m@.
 newtype Handler sig m =
   Handler { runHandler :: forall s x . sig (InterpretC s sig m) x -> InterpretC s sig m x }
 
@@ -34,26 +35,22 @@ class Reifies s a | s -> a where
 data Skolem
 
 
+-- | @Magic@ captures the GHC implementation detail of how single method type classes are implemented.
 newtype Magic a r =
   Magic (Reifies Skolem a => Tagged Skolem r)
 
 
+-- For more information on this technique, see the @reflection@ library. We use the formulation described in https://github.com/ekmett/reflection/issues/31 for better inlining.
+--
+-- Essentially we can view @k@ as internally a function of type @Reifies s a -> Tagged s r@, whch we can again view as just @a -> Tagged s r@ through @unsafeCoerce@. After this coercion, we just apply the function to @a@.
 reify :: forall a r . a -> (forall s . Reifies s a => Tagged s r) -> r
 reify a k =
   unsafeCoerce (Magic @a k) a
 
 
-instance (HFunctor eff, HFunctor sig, Reifies s (Handler eff m), Monad m, Carrier sig m) => Carrier (eff :+: sig) (InterpretC s eff m) where
-  eff (L eff) =
-    runHandler (unTag (reflect @s)) eff
-  eff (R other) =
-    InterpretC (eff (handleCoercible other))
-
-
-
 -- | Interpret an effect using a higher-order function.
 --
---   This involves a great deal less boilerplate than defining a custom 'Carrier' instance, at the expense of somewhat less performance. It’s a reasonable starting point for new interpretations, and if more performance or flexibility is required, it’s straightforward to “graduate” by replacing the relevant 'runInterpret' handlers with specialized 'Carrier' instances for the effects.
+-- Note that due to the higher-rank type, you have to use either '$' or explicit application when applying this interpreter. That is, you will need to write @runInterpret f (runInterpret g myPrgram)@ or @runInterpret f $ runInterpret g $ myProgram@. If you try and write @runInterpret f . runInterpret g@, you will unfortunately get a rather scary type error!
 --
 --   prop> run (runInterpret (\ op -> case op of { Get k -> k a ; Put _ k -> k }) get) === a
 runInterpret
@@ -81,8 +78,6 @@ runInterpret f m =
 
 -- | Interpret an effect using a higher-order function with some state variable.
 --
---   This involves a great deal less boilerplate than defining a custom 'Carrier' instance, at the expense of somewhat less performance. It’s a reasonable starting point for new interpretations, and if more performance or flexibility is required, it’s straightforward to “graduate” by replacing the relevant 'runInterpretState' handlers with specialized 'Carrier' instances for the effects.
---
 --   prop> run (runInterpretState (\ s op -> case op of { Get k -> runState s (k s) ; Put s' k -> runState s' k }) a get) === a
 runInterpretState
   :: (HFunctor eff, Monad m)
@@ -104,6 +99,13 @@ newtype InterpretC s (sig :: (* -> *) -> * -> *) m a =
 
 instance MonadTrans (InterpretC s sig) where
   lift = InterpretC
+
+
+instance (HFunctor eff, HFunctor sig, Reifies s (Handler eff m), Monad m, Carrier sig m) => Carrier (eff :+: sig) (InterpretC s eff m) where
+  eff (L eff) =
+    runHandler (unTag (reflect @s)) eff
+  eff (R other) =
+    InterpretC (eff (handleCoercible other))
 
 
 -- $setup