@ocharles 2019-07-17
Thanks for the extensive feedback! I admit I just dumped it into a file and forgot about it, I wasn't quite intending it to be in a mergeable state, but more as a conversation point. But I guess there's not much more to do! Let me blast through all the stylistic stuff, and we'll see what we're left with.