@ocharles 2019-07-17
I found the Cabal oddity - your `cabal.project` specifies `-O0`, which is overriding anything I put in the Cabal file. Maybe we drop this?